Simplify color calculation
We can simplify calculating the closest point on the edge of the RGB
gamut triangle by observing that the desired result is either at the
point closest to the desired result, or on the line between the two
closest points.

Thus, simplify the calculation by first determining the distance to all
three gamut points, then discarding the one that's farthest away,
and only considering the single line between the two remaining points.